requirements.txt
django==1.11.7
django-crispy-forms
django-import-export
django-reversion
django-formtools
djangorestframework-jwt
djangorestframework
django-cors-headers
django-filter
future
httplib2
six
xlwt
xlsxwriter
django
pillow
mysqlclient
markdown
httplib2
six
requests
coreapi
xlwt
xlsxwriter
drf-extensions
django-redis
pycryptodome
raven